requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
jsonpickle==2.0.0
multidict==5.2.0
numpy==1.22.0
psycopg2==2.9.1
sqlparse==0.4.2
tabulate==0.8.9
typing-extensions==3.10.0.2
yarl==1.7.0
yoyo-migrations==7.3.2
scikit-learn==0.24.2
pytz==2022.1
sentence-transformers==2.1.0
